Fix build with Dovecot 2.29+
This fixes the build for 2.29 (potentially) and 2.3 (version I tested
with). This, however, will most probably break builds for Dovecot <2.29,
although I'm not sure if it's desired to preserve compatibility with
these older versions. If it is, I'd be happy to try to change this, with
some guidance though, since my knowledge of C is quite limited.